dashboard/requirements.txt
pyserial==3.5
psycopg2-binary==2.9.9
